refactor(product list): split server code from client code
