+++ frontend/pages/Parts/index.tsx
@@ -1,6 +1,6 @@
 import { serverSidePropsWrapper } from '@helpers/next-helpers';
 import { ProductListType } from '@models/product-list';
-import { getProductListServerSideProps } from '@templates/product-list';
+import { getProductListServerSideProps } from '@templates/product-list/server';
 
 export { default } from '@templates/product-list';
